chatbot-service/requirements.txt
flask
flask-cors
python-dotenv
nltk
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
googletrans==4.0.0-rc1
sentence-transformers==2.2.2
transformers==4.30.2
tensorflow==2.10.1
huggingface_hub==0.14.1
textblob
datasets
torch
tqdm
symspellpy
faster-whisper
deep-translator
langdetect
pydub
gTTS
